bug-fix( user endpoints  )

+++ app/Http/Controllers/API/UserController.php
@@ -8,6 +8,7 @@ use App\Http\Controllers\Controller;
 use App\Http\Requests\StoreUserRequest;
 use App\Http\Requests\UpdateUserRequest;
 use App\Http\Resources\UserResource;
+use Illuminate\Http\JsonResponse;
 
 class UserController extends Controller
 {
@@ -19,86 +20,83 @@ class UserController extends Controller
     /**
      * Display a listing of the resource.
      */
-    public function index()
+    public function index(): JsonResponse
     {
         try {
             $users = User::all();
-            $args['data'] = UserResource::collection($users);
-            return $this->successResponse($args , 200 );
+            return $this->successResponse(UserResource::collection($users) );
         } catch (\Throwable $th) {
-            return $this->FailResponse($th->getMessage());
+            return $this->generalFailureResponse($th->getMessage());
         }
     }
 
     /**
      * Store a newly created resource in storage.
      */
-    public function store(StoreUserRequest $request)
+    public function store(StoreUserRequest $request): JsonResponse
     {
         try {
-            $validate = $request->validated();
-            $user= User::create([
+            User::create([
                 'name'    =>$request->name,
                 'email'   =>$request->email,
                 'password'=>$request->password,
                 'type'    =>$request->type,
             ]);
-            $args['message'] = 'user stored successfully ';
-            $args['data'] = new UserResource($user);
-            return $this->successResponse($args , 200 );
+            return $this->successOperationResponse('user stored successfully ');
         } catch (\Throwable $th) {
-            return $this->FailResponse($th->getMessage());
+            return $this->generalFailureResponse($th->getMessage());
         }
     }
 
     /**
      * Display the specified resource.
      */
-    public function show(string $id)
+    public function show(string $id): JsonResponse
     {
         try {
-            $user = User::findOrFail($id);
-            $args['data'] = new UserResource($user);
-            return $this->successResponse($args , 200 );
+            $user = User::find($id);
+            if (!$user)
+                return $this->errorResponse('there are no user for this id ');
+            return $this->successResponse(new UserResource($user) );
         } catch (\Throwable $th) {
-            return $this->FailResponse($th->getMessage());
+            return $this->generalFailureResponse($th->getMessage());
         }
     }
 
     /**
      * Update the specified resource in storage.
      */
-    public function update(UpdateUserRequest $request, string $id)
+    public function update(UpdateUserRequest $request, string $id): JsonResponse
     {
         try {
-            $validate = $request->validated();
-            $user = User::findOrFail($id);
+            $user = User::find($id);
+            if (!$user)
+                return $this->errorResponse('there are no user for this id ');
             $user->update([
                 'name'    =>$request->name    ??$user->name,
                 'email'   =>$request->email   ??$user->email,
                 'password'=>$request->password??$user->password,
                 'type'    =>$request->type    ??$user->type,
             ]);
-            $args['message'] = 'user updated successfully ';
-            $args['data'] = new UserResource($user);
-            return $this->successResponse($args , 200 );
+            return $this->successOperationResponse('user updated successfully ');
         } catch (\Throwable $th) {
-            return $this->FailResponse($th->getMessage());
+            return $this->generalFailureResponse($th->getMessage());
         }
     }
 
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(string $id)
+    public function destroy(string $id): JsonResponse
     {
         try {
-            $user = User::findOrFail($id);
+            $user = User::find($id);
+            if (!$user)
+                return $this->errorResponse('there are no user for this id ');
             $user->delete();
-            $args['message'] = 'user  deleted successfully ';
-            return $this->successResponse($args , 200);
+            return $this->successOperationResponse('user  deleted successfully ');
         } catch (\Throwable $th) {
-            return $this->FailResponse($th->getMessage());
+            return $this->generalFailureResponse($th->getMessage());
         }
     }
 }
